Updated code

+++ currencyNBP-API/src/main/java/com/currencyAPI/currencyNBPAPI/webClientNBP/dto/OpenGoldGoldDto.java
@@ -8,5 +8,5 @@ import java.time.LocalDate;
 public class OpenGoldGoldDto {
 
 //    private String data;
-    private int cena;
+    private float cena;
 }
